preparing frontend project

+++ roboshop/components/frontend.sh
@@ -4,9 +4,13 @@
 echo -e "\e[31mInstall NGINX package\e[0m"
 yum install nginx -y
 
-echo -e "\e[33mDownload frontend files\e[0m"
-curl -s -L -o /tmp/frontend.zip "https://github.com/roboshop-devops-project/frontend/archive/main.zip"
 
+echo -e "\e[33mDownload frontend files\e[0m"
+curl -s -L -o /tmp/frontend.zip "https://github.com/roboshop-devops-project/frontend/archive/main.zp"
+if [$? -ne 0]; then
+  echo download failed
+  exit
+fi
 echo -e "\e[33mCleanup files\e[0m"
 rm -rf /usr/share/nginx/html/
 cd /usr/share/nginx/html
